requirements.txt
django
psycopg2-binary
django-model-utils
celery
redis